requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
addict==0.4.0             # via cfgstack
astroid==1.6.5            # via pylint
backports.functools-lru-cache==1.5  # via astroid, pylint
cfgstack==0.1.post25
clip.py==0.3.2
configparser==3.5.0       # via pylint
contextlib2==0.5.4        # via raven
enum34==1.1.6             # via astroid
futures==3.2.0            # via isort
isort==4.3.4              # via pylint
lazy-object-proxy==1.3.1  # via astroid
logtool==0.3.post13
mccabe==0.6.1             # via pylint
path.py==8.1.2
pillow==5.2.0             # via reportlab
pylint==1.9.2             # via setuptools-lint
pyyaml==3.13
raven==5.32.0
reportlab==3.5.0
setuptools-lint==0.5.2    # via cfgstack
singledispatch==3.4.0.3   # via astroid, pylint
six==1.10.0               # via astroid, logtool, pylint, singledispatch
toml==0.9.4
wrapt==1.10.6             # via astroid, logtool
